backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.15
python-dotenv==1.0.0
motor==3.3.2
httpx==0.25.2
python-decouple==3.8